requirements.txt
aistudio-sdk==0.3.7
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
bce-python-sdk==0.9.45
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
distro==1.9.0
easyocr==1.7.2
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
lazy_loader==0.4
MarkupSafe==3.0.2
modelscope==1.29.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
openai==1.99.9
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
playwright==1.55.0
prettytable==3.16.0
protobuf==6.30.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tifffile==2025.9.9
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==10.4
zipp==3.23.0
